chore: refactor gateway-core and remove caching bits that should be generic and independent of the specific runtime
