+++ engine/crates/gateway-core/Cargo.toml
@@ -30,11 +30,10 @@ log = { path = "../log" }
 mediatype = "0.19"
 mime = "0.3"
 multipart-stream = "0.1"
-runtime = { workspace = true }
+runtime = { workspace = true, features = ["test-utils"] }
 serde_json = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
 
 [dev-dependencies]
-serde = { workspace = true }
 tokio = { workspace = true, features = ["macros"] }
